refactor: unit 함수 분리, 함수 이름 수정
